fix: move files into staff folder
