+++ pages/course/staff/editor/content/[id].tsx
